Several small optimizations (#8)
* Fix scrolling

I need to commit this to see if it actually works, but right now clicking the links does nothing because it has two # signs which does not indicate the ID on the page to link to as it should.

* Indicate need for quotes

When putting this as an option you can't just put the raw string, you need to indicate to the program that it is a string with quotes. Made this more clear to users.

* Stop example displaying as hyperlink

One of the example links was displayed as an actual link while the rest were just displayed as the URLs, so I fixed it to where they are all displayed uniformly as non-functioning examples.